feat: refactor auction maintenance components and introduce new CreateAuction and UpdateAuction functionalities
- Updated import paths for CreateAuction and UpdateAuction components to reflect new directory structure.
- Introduced CreateAuctionSuccess and UpdateAuctionSuccess components to provide user feedback upon successful auction creation and updates.
- Enhanced form handling and validation in CreateAuction and UpdateAuction components, improving user experience during auction setup and editing.
- Implemented dynamic routing for auction success pages, streamlining navigation after auction actions.

These changes collectively enhance the auction maintenance workflow, providing clearer feedback and improved user interaction during auction creation and editing processes.
